+++ src/fetchers/offer.fetcher.js
@@ -6,7 +6,13 @@ import { authTokenName } from "@/lib/authHelpers";
 
 export const fetchOfferList = async (page = 0, limit = 6) => {
     try {
-        const { data } = await axiosPrivate.get(`${API.offerList}/?type=vc&page=${page}&limit=${limit}`);
+        const { data } = await axiosPrivate.get(API.offerList, {
+            params: {
+                page,
+                limit,
+            }
+        });
+
         return data;
     } catch (e) {
         if (e?.status && e.status !== 401) {
@@ -33,7 +39,11 @@ export const fetchOfferProgress = async (offerId) => {
     if (!offerId) return {};
 
     try {
-        const { data } = await axiosPrivate.get(`${API.offerProgress}?offerId=${offerId}`);
+        const { data } = await axiosPrivate.get(API.offerProgress, {
+            params: {
+                offerId,
+            }
+        });
         return data;
     } catch (e) {
         if (e?.status && e.status !== 401) {
