+++ src/pages/app/offer/index.js
@@ -2,6 +2,7 @@ import { dehydrate, useQuery, useInfiniteQuery } from "@tanstack/react-query";
 
 import { processServerSideData } from "@/lib/serverSideHelpers";
 import { queryClient } from "@/lib/queryCache";
+import { cacheOptions } from "@/v2/helpers/query";
 import { fetchOfferList, fetchOfferStats } from "@/fetchers/offer.fetcher";
 import { AppLayout } from "@/v2/components/Layout";
 import { Metadata } from "@/v2/components/Layout";
@@ -18,27 +19,23 @@ export default function AppOpportunities({ session }) {
     const { tenantId: TENANT_ID, partnerId: PARTNER_ID } = session;
 
     const { data: statsList, isLoading: isStatsLoading, isError: isStatsError } = useQuery({
-        queryKey: ["offerStats", { TENANT_ID, PARTNER_ID }],
+        queryKey: ["offerStats", TENANT_ID, PARTNER_ID],
         queryFn: fetchOfferStats,
-        cacheTime: 30 * 60 * 1000,
-        staleTime: 15 * 60 * 1000,
-        refetchOnMount: false,
-        refetchOnWindowFocus: false,
+        ...cacheOptions,
     });
 
-    const { data: offersList, isLoading: isOffersLoading, isError: isOffersError, isFetchingNextPage, hasNextPage, fetchNextPage } = useInfiniteQuery({
-        queryKey: ["offerList", { TENANT_ID, PARTNER_ID }],
+    const { data: offersList, isLoading: isOffersLoading, isError: isOffersError, error, isFetchingNextPage, hasNextPage, fetchNextPage } = useInfiniteQuery({
+        queryKey: ["offerList", TENANT_ID, PARTNER_ID],
         queryFn: fetchOffers,
         getNextPageParam: (lastPage) => lastPage.nextPage,
         initialPageParam: 0,
-        cacheTime: 30 * 60 * 1000,
-        staleTime: 15 * 60 * 1000,
-        refetchOnMount: false,
-        refetchOnWindowFocus: false,
+        ...cacheOptions,
     });
 
     const offers = offersList?.pages.flatMap(page => page.offers) ?? [];
     const stats = statsList ?? {};
+
+    console.log('offers', offersList)
 
     if (isOffersLoading || isStatsLoading) {
         return (
@@ -49,7 +46,7 @@ export default function AppOpportunities({ session }) {
         );
     }
 
-    if (offers.length === 0 || isOffersError || isStatsError) {
+    if (!offers.length || isOffersError || isStatsError) {
         return (
             <div className="col-span-12 max-h-[40vh]">
                 <Metadata title="Opportunities" />
@@ -64,8 +61,8 @@ export default function AppOpportunities({ session }) {
 export const getServerSideProps = async ({ req, res }) => {
     const customLogicCallback = async (account) => {
         const { tenantId: TENANT_ID, partnerId: PARTNER_ID } = account;
-        await queryClient.prefetchQuery(["offerList", { TENANT_ID, PARTNER_ID }], fetchOfferList);
-        await queryClient.prefetchQuery(["offerStats", { TENANT_ID, PARTNER_ID }], fetchOfferStats);
+        await queryClient.prefetchQuery(["offerList", TENANT_ID, PARTNER_ID], fetchOfferList);
+        await queryClient.prefetchQuery(["offerStats", TENANT_ID, PARTNER_ID], fetchOfferStats);
         
         return {
             additionalProps: {
